Extract composer index tabs UI
