+++ assets/i18n/languages.json
@@ -1,7 +1,7 @@
 [
-  { "value": "en", "label": "English", "module": "./en.js?v=press-system-v3.4.26" },
-  { "value": "chs", "label": "简体中文", "module": "./chs.js?v=press-system-v3.4.26" },
-  { "value": "cht-tw", "label": "正體中文（台灣）", "module": "./cht-tw.js?v=press-system-v3.4.26" },
-  { "value": "cht-hk", "label": "繁體中文（香港）", "module": "./cht-hk.js?v=press-system-v3.4.26" },
-  { "value": "ja", "label": "日本語", "module": "./ja.js?v=press-system-v3.4.26" }
+  { "value": "en", "label": "English", "module": "./en.js?v=press-system-v3.4.27" },
+  { "value": "chs", "label": "简体中文", "module": "./chs.js?v=press-system-v3.4.27" },
+  { "value": "cht-tw", "label": "正體中文（台灣）", "module": "./cht-tw.js?v=press-system-v3.4.27" },
+  { "value": "cht-hk", "label": "繁體中文（香港）", "module": "./cht-hk.js?v=press-system-v3.4.27" },
+  { "value": "ja", "label": "日本語", "module": "./ja.js?v=press-system-v3.4.27" }
 ]
