+++ assets/js/composer-content-staging.js
@@ -1,8 +1,8 @@
-import { createCommitFileCollector } from './composer-staging.js?v=press-system-v3.4.26';
+import { createCommitFileCollector } from './composer-staging.js?v=press-system-v3.4.27';
 import {
   listLocalMarkdownAssetReferences,
   planManagedContentDeletions
-} from './repository-deletions.js?v=press-system-v3.4.26';
+} from './repository-deletions.js?v=press-system-v3.4.27';
 
 export function createContentCommitStagingProvider({
   getDynamicEditorTabs = () => new Map(),
