+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.26",
-  "tag": "v3.4.26",
+  "version": "3.4.27",
+  "tag": "v3.4.27",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.25 <3.4.26"
+      ">=3.4.26 <3.4.27"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.25 first."
+    "message": "Update to v3.4.26 first."
   }
 }
