+++ index_editor_preview.html
@@ -13,6 +13,6 @@
 </head>
 <body>
   <div class="editor-preview-status" id="editorPreviewStatus">Loading preview...</div>
-  <script type="module" src="assets/js/editor-preview-runtime.js?v=press-system-v3.4.26"></script>
+  <script type="module" src="assets/js/editor-preview-runtime.js?v=press-system-v3.4.27"></script>
 </body>
 </html>
